# Conflicts:
#	src/main/java/org/promptunit/dsl/PromptResultAssert.java
#	src/main/java/org/promptunit/providers/anthropic/AnthropicEngine.java
#	src/main/java/org/promptunit/providers/llama/OllamaEngine.java
#	src/main/java/org/promptunit/providers/openai/OpenAIEngine.java
#	src/test/java/org/promptunit/examples/OpenAIExamplesIT.java

+++ src/main/java/org/promptunit/core/PromptInstance.java
@@ -14,8 +14,8 @@ public record PromptInstance(
 		List<Message> conversation,
 		String model,
 		String provider,
-		Float temperature,
-		Float topP,
+		Double temperature,
+		Double topP,
 		Integer maxTokens,
 		Optional<OutputSchema> outputSchema
 ) {
@@ -60,8 +60,8 @@ public record PromptInstance(
 		private final List<Message> conversation = new ArrayList<>();
 		private String model;
 		private String provider;
-		private Float temperature;
-		private Float topP;
+		private Double temperature;
+		private Double topP;
 		private Integer maxTokens;
 		private OutputSchema outputSchema;
 
@@ -90,12 +90,12 @@ public record PromptInstance(
 			return this;
 		}
 
-		public Builder withTemperature(Float temperature) {
+		public Builder withTemperature(Double temperature) {
 			this.temperature = temperature;
 			return this;
 		}
 
-		public Builder withTopP(Float topP) {
+		public Builder withTopP(Double topP) {
 			this.topP = topP;
 			return this;
 		}
